export MustShimLegacyState for resource tests
We also need to convert legacy states for helper resource tests.
