+++ terraform/context.go
@@ -1,6 +1,7 @@
 package terraform
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"log"
@@ -18,6 +19,7 @@ import (
 	"github.com/hashicorp/terraform/providers"
 	"github.com/hashicorp/terraform/provisioners"
 	"github.com/hashicorp/terraform/states"
+	"github.com/hashicorp/terraform/states/statefile"
 	"github.com/hashicorp/terraform/tfdiags"
 )
 
@@ -896,4 +898,38 @@ func parseVariableAsHCL(name string, input string, targetType config.VariableTyp
 	default:
 		panic(fmt.Errorf("unknown type %s", targetType.Printable()))
 	}
+}
+
+// shimLegacyState is a helper that takes the legacy state type and
+// converts it to the new state type.
+//
+// This is implemented as a state file upgrade, so it will not preserve
+// parts of the state structure that are not included in a serialized state,
+// such as the resolved results of any local values, outputs in non-root
+// modules, etc.
+func shimLegacyState(legacy *State) (*states.State, error) {
+	if legacy == nil {
+		return nil, nil
+	}
+	var buf bytes.Buffer
+	err := WriteState(legacy, &buf)
+	if err != nil {
+		return nil, err
+	}
+	f, err := statefile.Read(&buf)
+	if err != nil {
+		return nil, err
+	}
+	return f.State, err
+}
+
+// MustShimLegacyState is a wrapper around ShimLegacyState that panics if
+// the conversion does not succeed. This is primarily intended for tests where
+// the given legacy state is an object constructed within the test.
+func MustShimLegacyState(legacy *State) *states.State {
+	ret, err := shimLegacyState(legacy)
+	if err != nil {
+		panic(err)
+	}
+	return ret
 }
